Move some setup around
